requirements.txt
pyyaml
gunicorn==20.1.0
flask